Fix error catching.

+++ main.py
@@ -7,26 +7,18 @@ import random
 def get_picture_data(picture_id):
     url = f'http://xkcd.com/{picture_id}/info.0.json'
     response = requests.get(url)
-    try:
-        response.raise_for_status()
-        return response.json()
-    except requests.exceptions.HTTPError:
-        print('Picture not found')
-        exit()
+    response.raise_for_status()
+    return response.json()
 
 
 def download_picture(picture_id):
     picture_link = get_picture_data(picture_id)['img']
 
     picture_response = requests.get(picture_link)
-    try:
-        picture_response.raise_for_status()
-        filename = f'{picture_id}.png'
-        with open(filename, 'wb') as image:
-            image.write(picture_response.content)
-    except requests.exceptions.HTTPError:
-        print('Picture not found')
-        exit()
+    picture_response.raise_for_status()
+    filename = f'{picture_id}.png'
+    with open(filename, 'wb') as image:
+        image.write(picture_response.content)
 
 
 def get_info_for_upload(group_id, *args):
@@ -37,11 +29,9 @@ def get_info_for_upload(group_id, *args):
     }
     url = f'https://api.vk.com/method/photos.getWallUploadServer'
     response = requests.get(url, params=params).json()
-    try:
-        print(response['error']['error_msg'])
-        exit()
-    except KeyError:
-        return response['response'][args[0]]
+
+    if 'error' in response:
+        raise requests.exceptions.HTTPError(response['error']['error_msg'])
 
 
 def upload_picture(pic_num, group_id):
@@ -50,12 +40,12 @@ def upload_picture(pic_num, group_id):
     pic_file = open(f'{pic_num}.png', 'rb')
     files = {'photo': pic_file}
     response = requests.post(upload_url, files=files).json()
+
     pic_file.close()
-    try:
-        print(response['error']['error_msg'])
-        exit()
-    except KeyError:
-        return response['server'], response['photo'], response['hash']
+    if 'error' in response:
+        raise requests.exceptions.HTTPError(response['error']['error_msg'])
+
+    return response['server'], response['photo'], response['hash']
 
 
 def get_data_for_post(pic_num, group_id):
@@ -70,11 +60,11 @@ def get_data_for_post(pic_num, group_id):
     }
     url = f'https://api.vk.com/method/photos.saveWallPhoto'
     response = requests.post(url, params=params).json()
-    try:
-        print(response['error']['error_msg'])
-        exit()
-    except KeyError:
-        return response['response'][0]['id'], response['response'][0]['owner_id']
+
+    if 'error' in response:
+        raise requests.exceptions.HTTPError(response['error']['error_msg'])
+
+    return response['response'][0]['id'], response['response'][0]['owner_id']
 
 
 def post_picture(pic_num, group_id):
@@ -89,21 +79,19 @@ def post_picture(pic_num, group_id):
     }
     url = f'https://api.vk.com/method/wall.post'
     response = requests.get(url, params=params).json()
-    try:
-        print(response['error']['error_msg'])
-        exit()
-    except KeyError:
-        os.remove(f'{pic_num}.png')
+
+    if 'error' in response:
+        raise requests.exceptions.HTTPError(response['error']['error_msg'])
+
+    os.remove(f'{pic_num}.png')
 
 
 def get_last_picture_num():
     url = 'https://xkcd.com/info.0.json'
     response = requests.get(url)
-    try:
-        response.raise_for_status()
-        return response.json()['num']
-    except requests.exceptions.HTTPError:
-        exit()
+
+    response.raise_for_status()
+    return response.json()['num']
 
 
 if __name__ == '__main__':
